Added can? verification to render or not actions New/Edit/Destroy

+++ lib/active_admin/resource/action_items.rb
@@ -48,21 +48,21 @@ module ActiveAdmin
       def add_default_action_items
         # New Link on all actions except :new and :show
         add_action_item :except => [:new, :show] do
-          if controller.action_methods.include?('new')
+          if controller.action_methods.include?('new') && can?(:create, active_admin_config.resource_name.constantize)
             link_to(I18n.t('active_admin.new_model', :model => active_admin_config.resource_name), new_resource_path, :class => "active_admin new_action")
           end
         end
 
         # Edit link on show
         add_action_item :only => :show do
-          if controller.action_methods.include?('edit')
+          if controller.action_methods.include?('edit') && can?(:update, active_admin_config.resource_name.constantize)
             link_to(I18n.t('active_admin.edit_model', :model => active_admin_config.resource_name), edit_resource_path(resource), :class => "active_admin edit_action")
           end
         end
 
         # Destroy link on show
         add_action_item :only => :show do
-          if controller.action_methods.include?("destroy")
+          if controller.action_methods.include?("destroy") && can?(:destroy, active_admin_config.resource_name.constantize)
             link_to(I18n.t('active_admin.delete_model', :model => active_admin_config.resource_name),
               resource_path(resource), :class => "active_admin delete_action",
               :method => :delete, :confirm => I18n.t('active_admin.delete_confirmation'))
